Fix cachestats() LLP64 truncation and mislabeled cold-read test
Use %llu with unsigned long long for all four cachestats() fields to
avoid truncation on platforms where size_t is 64-bit but unsigned long
is 32-bit (Windows LLP64).

Rewrite TC002 to test genuine cold-miss behavior: create a fresh object
and read an attribute that was never in the cache. Old TC002 only proved
re-read hits because write-through already warmed the cache. Renumber
TC003 for cached re-read validation, TC004 for timed benchmark.

+++ mux/modules/engine/functions.cpp
@@ -11197,11 +11197,11 @@ static FUNCTION(fun_cachestats)
     cache_get_stats(&st);
 
     safe_tprintf_str(buff, bufc,
-        T("%llu %llu %lu %lu"),
+        T("%llu %llu %llu %llu"),
         static_cast<unsigned long long>(st.hits),
         static_cast<unsigned long long>(st.misses),
-        static_cast<unsigned long>(st.entries),
-        static_cast<unsigned long>(st.size));
+        static_cast<unsigned long long>(st.entries),
+        static_cast<unsigned long long>(st.size));
 }
 
 static FUNCTION(fun_config)
